+++ crates/rdlp-extractor/src/extractors/tnaflix/mod.rs
@@ -13,10 +13,13 @@
 //! - `search` - TNAFlix search extractor (`TNAFlixSearchExtractor`)
 //! - `tnaflix_search_helpers` - HTML search result parsing for TNAFlix
 //! - `search_patterns` - Search URL builders and filter descriptors
+//! - `empflix_search` - EMPFlix search extractor (`EMPFlixSearchExtractor`)
+//! - `moviefap_search` - MovieFap search extractor (`MovieFapSearchExtractor`)
 
 mod ajax;
 mod empflix_search;
 mod info;
+mod moviefap_search;
 mod moviefap_search_helpers;
 mod moviefap_search_patterns;
 mod patterns;
@@ -26,152 +29,13 @@ mod tnaflix_search_helpers;
 
 pub use empflix_search::EMPFlixSearchExtractor;
 pub use info::TNAFlixExtractor;
+pub use moviefap_search::MovieFapSearchExtractor;
 pub use search::TNAFlixSearchExtractor;
-
-use async_trait::async_trait;
-use log::debug;
-use rdlp_core::{ExtractionContext, Result};
-use std::time::Duration;
 
 #[cfg(test)]
 use crate::base::common::BaseExtractor;
 #[cfg(test)]
 use rdlp_core::SearchExtractor;
-
-/// Rate limit delay between search page fetches (500 ms)
-const PAGE_RATE_LIMIT_MS: u64 = 500;
-
-/// MovieFap search extractor
-///
-/// Parses MovieFap's distinct HTML search result structure.
-/// Uses `moviefap_search` and `moviefap_search_patterns` internally.
-pub struct MovieFapSearchExtractor;
-
-impl MovieFapSearchExtractor {
-    /// Create a new MovieFap search extractor.
-    #[must_use]
-    pub fn new() -> Self {
-        Self
-    }
-
-    /// Fetch a single search results page and return `(results, max_page_number)`.
-    async fn fetch_search_page(
-        &self,
-        query: &rdlp_types::SearchQuery,
-        page: usize,
-        ctx: &ExtractionContext,
-    ) -> Result<(Vec<rdlp_types::SearchResultPreview>, usize)> {
-        let page_url = moviefap_search_patterns::build_search_url(query, page);
-        debug!(page; "[MovieFap] Fetching search page: {}", rdlp_security::sanitize_for_logging(&page_url));
-
-        let webpage = crate::base::common::BaseExtractor::fetch_webpage(&page_url, ctx).await?;
-        let page_results = moviefap_search_helpers::parse_search_results(&webpage);
-        let max_pages = moviefap_search_helpers::parse_pagination(&webpage).unwrap_or(1);
-
-        debug!(
-            count = page_results.len(),
-            max_pages;
-            "[MovieFap] Search page {page} returned {} results",
-            page_results.len()
-        );
-
-        Ok((page_results, max_pages))
-    }
-
-    /// Collect all pages up to `MAX_PLAYLIST_SIZE` results.
-    async fn search_all_pages(
-        &self,
-        query: &rdlp_types::SearchQuery,
-        ctx: &ExtractionContext,
-    ) -> Result<Vec<rdlp_types::SearchResultPreview>> {
-        moviefap_search_helpers::validate_search_filters(&query.filters)?;
-
-        let max_results = query
-            .max_results
-            .unwrap_or(crate::base::common::MAX_PLAYLIST_SIZE);
-
-        let mut all_results = Vec::new();
-        let mut page = 1usize;
-
-        loop {
-            let (page_results, max_pages) = match self.fetch_search_page(query, page, ctx).await {
-                Ok(r) => r,
-                Err(e) => {
-                    debug!(page; "[MovieFap] Failed to fetch search page, returning partial results: {e}");
-                    break;
-                }
-            };
-
-            if page_results.is_empty() {
-                debug!(page; "[MovieFap] No results on page, stopping pagination");
-                break;
-            }
-
-            all_results.extend(page_results);
-
-            if all_results.len() >= max_results {
-                all_results.truncate(max_results);
-                break;
-            }
-
-            if page >= max_pages {
-                break;
-            }
-
-            page += 1;
-            tokio::time::sleep(Duration::from_millis(PAGE_RATE_LIMIT_MS)).await;
-        }
-
-        debug!(count = all_results.len(), pages = page; "[MovieFap] Search complete");
-
-        Ok(all_results)
-    }
-}
-
-impl Default for MovieFapSearchExtractor {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[async_trait]
-impl rdlp_core::SearchExtractor for MovieFapSearchExtractor {
-    fn name(&self) -> &str {
-        "MovieFap"
-    }
-
-    fn supported_filters(&self) -> Vec<rdlp_types::SearchFilterDescriptor> {
-        moviefap_search_patterns::search_filter_descriptors()
-    }
-
-    async fn search(
-        &self,
-        query: &rdlp_types::SearchQuery,
-        ctx: &ExtractionContext,
-    ) -> Result<Vec<rdlp_types::SearchResultPreview>> {
-        self.search_all_pages(query, ctx).await
-    }
-
-    async fn search_page(
-        &self,
-        query: &rdlp_types::SearchQuery,
-        ctx: &ExtractionContext,
-    ) -> Result<rdlp_types::SearchPageResponse> {
-        moviefap_search_helpers::validate_search_filters(&query.filters)?;
-
-        let page = query.page.unwrap_or(1) as usize;
-        let (page_results, max_pages) = self.fetch_search_page(query, page, ctx).await?;
-
-        let has_more = page < max_pages && !page_results.is_empty();
-
-        Ok(rdlp_types::SearchPageResponse {
-            results: page_results,
-            page: page as u32,
-            has_more,
-            total_estimate: None,
-        })
-    }
-}
 
 #[cfg(test)]
 mod tests {
@@ -488,23 +352,23 @@ mod async_tests {
     /// Minimal no-op JsEngine for tests.
     struct NoOpJsEngine;
 
-    #[async_trait]
+    #[async_trait::async_trait]
     impl JsEngine for NoOpJsEngine {
-        async fn eval(&self, _code: &str) -> Result<serde_json::Value> {
+        async fn eval(&self, _code: &str) -> rdlp_core::Result<serde_json::Value> {
             Ok(serde_json::Value::Null)
         }
         async fn eval_with_context(
             &self,
             _code: &str,
             _context: &serde_json::Value,
-        ) -> Result<serde_json::Value> {
+        ) -> rdlp_core::Result<serde_json::Value> {
             Ok(serde_json::Value::Null)
         }
         async fn call_function(
             &self,
             _name: &str,
             _args: &[serde_json::Value],
-        ) -> Result<serde_json::Value> {
+        ) -> rdlp_core::Result<serde_json::Value> {
             Ok(serde_json::Value::Null)
         }
     }
@@ -512,18 +376,21 @@ mod async_tests {
     /// Minimal no-op CookieJar for tests.
     struct NoOpCookieJar;
 
-    #[async_trait]
+    #[async_trait::async_trait]
     impl CookieJar for NoOpCookieJar {
-        async fn cookies(&self, _url: &str) -> Result<Vec<String>> {
+        async fn cookies(&self, _url: &str) -> rdlp_core::Result<Vec<String>> {
             Ok(vec![])
         }
-        async fn add_cookie(&self, _url: &str, _cookie: &str) -> Result<()> {
+        async fn add_cookie(&self, _url: &str, _cookie: &str) -> rdlp_core::Result<()> {
             Ok(())
         }
-        async fn load_from_browser(&self, _browser: rdlp_types::BrowserType) -> Result<usize> {
+        async fn load_from_browser(
+            &self,
+            _browser: rdlp_types::BrowserType,
+        ) -> rdlp_core::Result<usize> {
             Ok(0)
         }
-        async fn load_from_file(&self, _path: &std::path::Path) -> Result<usize> {
+        async fn load_from_file(&self, _path: &std::path::Path) -> rdlp_core::Result<usize> {
             Ok(0)
         }
     }
